(tae): small code refactor 3
