+++ pkg/vm/engine/tae/db/tables_test.go
@@ -1,4 +1,4 @@
-package test
+package db
 
 import (
 	"bytes"
